requirements.txt
asn1crypto==0.24.0
astroid==1.6.5
autopep8==1.4
backports.functools-lru-cache==1.5
cffi==1.11.5
Click==7.0
configparser==3.5.0
coverage==4.5.1
cryptography==2.3.1
enum34==1.1.6
Flask==1.0.2
idna==2.7
ipaddress==1.0.22
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jwt==0.5.4
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
nose2==0.8.0
psycopg2-binary==2.7.5
pyasn1==0.4.4
pycodestyle==2.4.0
pycparser==2.19
pyj==0.0
PyJWT==1.4.0
pylint==1.9.3
singledispatch==3.4.0.3
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
